LU-19264 ldlm: server notify lock clients to reclaim LRU locks
In large-scale Lustre cluster with a number of clients, the server
side lock namespace must manage all locks granted to clients. And
these locks are cached in server side memory. Due to the limited
server memory capacity, the server must reclaim locks when it is
under memory pressure. However, when lock reclaim is triggered,
the LRU or frequency of locks cached on the server lock namespace
may not reflect recent used accurately.
This can lead to premature reclaim of locks that are actually
still in active use or may be reused in the near future by the
client. The lock revocation will drop the cache data or metadata
protected by the lock on the client. This may result in
unnecessary network traffic, re-reading the data (such as
memory resident library files) dropped from cache wrongly and
degraded system performance.

To solve this problem, this patch implements a mechanism to let
the lock server to notify clients, which have much accurate and
realtime information about lock usage, to make better decision for
lock reclaim on client side.

Signed-off-by: Yingjin Qian <qian@ddn.com>
Change-Id: I55a7d5876e5dfbdfbfc84d073e86dc0156f886cd